modified how variables like studyID are passed between scripts (.sh, R, parsed with docopts), changed as.integer to as.numeric, and added in Rscript to run R script instead of as a shell command as was set previously. With the version of the code in this commit, all 5 studies in study.csv were downloaded and validated

+++ db/validate_import.r
@@ -48,7 +48,7 @@ if(interactive()) {
   .eda <- as.logical(ag$eda)
   rd <- is_rstudio_project$make_fix_file(.script)
   
-  .studyid <- as.integer(ag$studyid)
+  .studyid <- as.numeric(ag$studyid)
   
   if(length(ag$db)==0) {
     .dbPF <- file.path(.wd,'data','mosey.db')
